+++ modules/core/dependency/python-ihm/make-release.sh
@@ -4,6 +4,7 @@
 # - Update AuditConformDumper to match latest IHM dictionary if necessary
 # - Run util/validate-outputs.py to make sure all example outputs validate
 #   (cd util; PYTHONPATH=.. python3 ./validate-outputs.py)
+# - Make sure all python-modelcif tests work using new IHM version
 # - Update ChangeLog.rst, util/debian/changelog, and util/python-ihm.spec
 #   with the release number and date
 # - Update release number in ihm/__init__.py, MANIFEST.in, and setup.py
@@ -25,3 +26,5 @@ rm -f "src/ihm_format_wrap_${VERSION}.c"
 
 echo "Now use 'twine upload dist/ihm-${VERSION}.tar.gz' to publish the release on PyPi."
 echo "Then, update the conda-forge, COPR, PPA, and Homebrew packages to match."
+echo "For COPR, use dist/ihm-${VERSION}.tar.gz together with util/python-ihm.spec"
+echo "For the PPA, use the GitHub release tarball, renamed to python-ihm_${VERSION}.orig.tar.gz"